clean up code
git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@11425 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ igb/src/com/affymetrix/igb/prefs/TrackDefaultView.java
@@ -393,7 +393,7 @@ public final class TrackDefaultView extends TrackPreferences implements ListSele
 								style.setTrackNameSize((Float) value);
 							}
 							style.setTrackNameSize((Float) value);
-							trackNameSizeComboBox.setSelectedItem((Float) value);
+							trackNameSizeComboBox.setSelectedItem(value);
 							break;
 						case COL_LABEL_FIELD:
 							if (!style.equals(default_annot_style)) {
